Fixed bugs and extended map
Fixed an issue where if the user clicked on a pixel on the map that is contained in more than one country, it would select all those countries at the same time. Added a few more countries to the map. Made a small program to help with the reversal of arrays of coordinates, since it is needed every so often.